Tested: PATH COMMAND CHAINING, Note Distance Decreased
After some hard work, we can finally chain OTF paths with other commands. The thing with OTF paths is that the are explicitly scheduled, which interrupts commands calling it. For example, putting the scheduler inside InstantCommand actually exits the command as interrupted, and like wise when putting scheduler inside a normal command type .java file, end() method is called because interrupt is true. As such, there really is no way to convert OTF path into a command, except to have event markers toggle boolean variables and let Robot Container receive triggers. Chaining here is done such that two booleans, one for determining when the path ends, and another for determining if NotePickUp was called, are factors to a trigger. When NotePickUp is true and pathActive is false (aka ended), then it triggers other commands (and this can recursively call NotePickUp again). This is the only way we can chain paths to other commands.

Since path planner is all about the robot center, the robot will move to the center of a detected note - meaning it will run over on top the note's location. The issue with this is that when picking up notes close to barriers or walls, the robot will definitely crash its intake. If the robot attempts to lift the intake when it is stuck, it will slip and damage the intake, making it unusable. To avoid this, we'll need to make sure that the robot DOES NOT fully run over on top the note, but rather the intake should. The intake to the center of the robot is roughly 0.5m, so we'll reduce the X distance of the note (forward distance relative to robot). When approaching notes directly in front, the robot tends to lean to the right by 15 cm, so I brought back the YOFFSET. Now, forward notes have no problem being picked up, but notes that span to the sides of the view are basically guaranteed to miss due to the robot not being able to reach the note. The solution for this is simply to turn first, then run the path separately. Additionally, for consistency reasons, we've reduced the maximum note distance acceptable for detection down to 2.5 meters rather than 3.

Shooter speed seems to peak at 70%.

Gear ratio for drive motors are some how back to being 6.7...

Paths now have three categories for pick up notes: really close means the robot will retreat before picking up; relatively close means the robot will approach the note slower, and normal distance is normal speed. Maximum acceleration and velocity decreased to 1.