Problematic Auto Turning, but pretty good when working
BTW, commands are initiated when the robot container is initiated; all values passed in to commands when constructed will stay that way, those values will not change. Commands are "run" when it is INITIALIZED. Any updating variables based on periodic activity should be under execute() and initialize(), not under the constructor.

THe AutoTurn command is having trouble because it is running into the +-180 barrier rotation range issues. The desired direction may often be outside the range, and short cut turning to it can cause problems...