Tested: Pretty Accurate Note Path!
Robot now turns to specified direction on the spot (currently only note path uses this), waits a bit, then runs a note path directly towards the note. The robot turning is to ensure 1. the note is within view, and 2. the note is directly in front to reduce

AutoTurn does the math to accommodate for passing +- 180 angle boundaries